+++ nofos/composer/templatetags/edit_mode_label.py
@@ -1,4 +1,4 @@
-from composer.utils import get_edit_mode_label
+from composer.utils import get_edit_mode_label, get_edit_mode_label_class
 from django import template
 
 register = template.Library()
@@ -11,3 +11,12 @@ def edit_mode_label(value: str) -> str:
     Usage: {{ subsection.edit_mode|edit_mode_label }}
     """
     return get_edit_mode_label(value)
+
+
+@register.filter
+def edit_mode_label_class(value: str) -> str:
+    """
+    Translate an edit_mode value into a class to apply to the label tag <span>
+    Usage: {{ subsection.edit_mode|edit_mode_label_class }}
+    """
+    return get_edit_mode_label_class(value)
