+++ nofos/composer/tests/test_utils.py
@@ -4,7 +4,7 @@ from django.test import SimpleTestCase
 
 class GetEditModeLabelTests(SimpleTestCase):
     def test_known_values(self):
-        self.assertEqual(get_edit_mode_label("variables"), "Variables")
+        self.assertEqual(get_edit_mode_label("variables"), "Certain text")
         self.assertEqual(get_edit_mode_label("yes_no"), "Yes/No")
         self.assertEqual(get_edit_mode_label("locked"), "Locked")
